requirements.txt
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
idna==2.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
urllib3==1.23
